+++ com.unity.formats.alembic/CHANGELOG.md
@@ -8,6 +8,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 - Prevent warnings related to FindObjectOfType<> calls on package installation with Unity Editor 2023.1+.
 
+### Removed
+- Removed support of Stadia as a build target.
+
 ## [2.3.2] - 2022-05-08
 ### Added
 - At import, additional data is added to GameObject cameras to conform with URP or HDRP rendering pipelines.